pyproject.toml
[project]
name = "swapi-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "dotenv>=0.9.9",
    "fastapi[standard-no-fastapi-cloud-cli]>=0.117.1",
    "orjson>=3.11.3",
    "pydantic-settings>=2.11.0",
    "sqlalchemy[asyncio]>=2.0.43",
]

[project.optional-dependencies]
dev = [
    "pre-commit>=4.3.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short --cov=src --cov-report=term-missing --cov-report=html"
asyncio_mode = "auto"

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\(.*ABC.*\\):",
    "@(abc\\.)?abstractmethod",
    "from .* import .*",
    "import .*"
]
